Extract code to modify rule sets
Extract the code to modify rule sets to a helper type.
